refactor: extract AddRepositoryCombobox hooks (#80)
Extract 3 custom hooks from AddRepositoryCombobox (12 useState → 2):
- useOrganizationData: user profile + org fetching
- useRepositoryData: repo fetching with org merge
- useRepositorySearch: search, selection, add state
